requirements.txt
Django==1.7.5
django-braces==1.4.0
django-cas-ng==3.4.2
django-crispy-forms==1.4.0
django-guardian==1.2.5
django-haystack==2.3.1
django-model-utils==2.2
django-randomslugfield==0.3.0
django-taggit==0.12.2
django-templated-email==0.4.9
djangorestframework==3.0.5
psycopg2==2.6
requests==2.6.0
six==1.9.0
wsgiref==0.1.2
